requirements.txt
# Core dependencies
PyGithub==2.8.1
python-dotenv==1.2.1
pyyaml==6.0.3

# Data processing
pandas==3.0.1
numpy==2.4.2

# Excel generation
openpyxl==3.1.5
xlsxwriter==3.2.9

# Utilities
aiohttp==3.13.3
requests==2.32.5
urllib3==2.6.3
colorama==0.4.6
pydantic-settings==2.13.1

# Testing
pytest==9.0.2
pytest-cov==7.0.0
pytest-mock==3.15.1

# Code quality
black==26.1.0
flake8==7.3.0
pylint==4.0.5
mypy==1.19.1